monitoring

+++ api/schemas.py
@@ -5,7 +5,7 @@ class RawInputData(BaseModel):
     UNIXTime: int = Field(..., example=1481298952)
     Data: str = Field(..., example="12/9/2016 12:00:00 AM")
     Time: str = Field(..., example="05:55:52")
-    Temperature: float = Field(..., example=48)
+    Temperature: int = Field(..., example=48)
     Pressure: float = Field(..., example=30.38)
     Humidity: int = Field(..., example=89)
     WindDirection_Degrees: float = Field(..., example=247.97)
